+++ src/command/restore/protocol.c
@@ -60,7 +60,7 @@ restoreFileProtocol(PackRead *const param, ProtocolServer *const server)
             file.blockIncrMapSize = pckReadU64P(param);
 
             if (file.blockIncrMapSize != 0)
-                file.blockIncrSize = pckReadU64P(param);
+                file.blockIncrSize = (size_t)pckReadU64P(param);
 
             file.manifestFile = pckReadStrP(param);
